requirements.txt
Flask==2.3.2
requests==2.31.0
openai==0.27.0
praw==7.7.0
python-dotenv==1.0.0
